refactor: 유틸리티 네이밍 변경
