+++ api/src/testFixtures/kotlin/com/retoday/api/snippet/CommonSnippets.kt
@@ -1,8 +1,8 @@
 package com.retoday.api.snippet
 
+import com.retoday.api.extension.desc
+import com.retoday.api.extension.fieldsOf
 import com.retoday.api.global.dto.ErrorResponse
-import com.retoday.api.util.desc
-import com.retoday.api.util.fieldsOf
 
 val errorResponseFields =
     fieldsOf(
